pyproject.toml
[build-system]
requires = ["setuptools >= 61.0.0", "versioneer[toml]==0.29"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]


[tool.setuptools.package-data]
jack_connection_manager = ["*.yml"]

# automatically handle version numbers
[tool.versioneer]
VCS = "git"
style = "pep440"
versionfile_source = "src/jack_connection_manager/_version.py"
versionfile_build = "jack_connection_manager/_version.py"
tag_prefix = ""
parentdir_prefix = "jack_connection_manager-"

[project]
name = "jack_connection_manager"
description = "A tool for automatically connecting jack clients with lots of ports"
authors = [{ name = "Max Weidauer", email = "weidauer@campus.tu-berlin.de" }]
readme = "README.md"
requires-python = ">=3.10"
dynamic = ["version"]
dependencies = ["JACK-Client", "click", "pyYAML", "sdnotify"]
license = { file = "LICENSE" }

[project.scripts]
jack-connection-manager = "jack_connection_manager.jack_connection_manager:main"

[project.urls]
Github = "https://github.com/tu-studio/jack-connection-manager"
Issues = "https://github.com/tu-studio/jack-connection-manager/issues"
